upload_school_data/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
pymysql>=1.0.0
python-dotenv>=1.0.0
